add createPackage command
